Lots of perf improvements with lazy functions

+++ bash-sources/rbenv.sh
@@ -1,47 +1,33 @@
 #!/usr/bin/env bash
 
-function install_ruby {
-    sd::log "Installing ruby 3.2.1 with rbenv"
-    rbenv install 3.2.1
-    rbenv global 3.2.1
-}
-sd::lazy_install_hook ruby install_ruby
-sd::lazy_install_hook irb install_ruby
+THIS_PROG="$BASH_SOURCE"
 
+function _ruby_irb_init {
+    if ! sd::bin_exists ruby ; then
+        rbenv install 3.2.1
+        rbenv global 3.2.1
+    fi
+    sd::log::debug should be inited
+}
+
+function check_rbenv {
+    [ -f ~/.rbenv/bin/rbenv ]
+}
 
 function install_rbenv {
-    sd::log "installing rbenv dependencies"
-    sd::log::command sudo apt-get install \
-        autoconf \
-        bison \
-        patch \
-        build-essential \
-        rustc \
-        libssl-dev \
-        libyaml-dev \
-        libreadline6-dev \
-        zlib1g-dev \
-        libgmp-dev \
-        libncurses5-dev \
-        libffi-dev \
-        libgdbm6 \
-        libgdbm-dev \
-        libdb-dev \
-        uuid-dev
-
-    sd::log "rbenv not found on path, installing it"
-    url="https://github.com/rbenv/rbenv-installer/raw/HEAD/bin/rbenv-installer"
-    curl -fsSL "$url" | bash
-    do_rbenv_init
+    git clone --depth 1 https://github.com/rbenv/rbenv.git ~/.rbenv
+    git clone --depth 1 https://github.com/rbenv/ruby-build.git ~/.rbenv/plugins/ruby-build
 }
-sd::lazy_install_hook --needs-sudo rbenv install_rbenv
 
-function do_rbenv_init {
-    local rbenv_init_commands=$(rbenv init - | grep -v "export PATH")
-    eval "$rbenv_init_commands"
-
+function init_rbenv {
+    to_eval=$(~/.rbenv/bin/rbenv init - bash | grep -v PATH)
     sd::path::prepend "${HOME}/.rbenv/bin"
     sd::path::prepend "${HOME}/.rbenv/shims"
+
+    eval "$to_eval"
 }
 
-sd::bin_exists rbenv && do_rbenv_init
+
+sd::func::lazy_create --func rbenv --install install_rbenv --check check_rbenv --init init_rbenv
+sd::func::jit ruby _ruby_irb_init
+sd::func::jit irb _ruby_irb_init
